pyproject.toml
[project]
name = "suagent"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.2",
    "duckduckgo-search>=8.1.1",
    "openai>=2.24.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
]
